setContext in [space]/+layout
