+++ src/lib/components/ChatMode.svelte
@@ -6,12 +6,10 @@
   import { page } from "$app/state";
   import { g } from "$lib/global.svelte";
 
-  import { setContext } from "svelte";
   import { slide } from "svelte/transition";
-  import type { Item } from "$lib/tiptap/editor";
   import { getProfile } from "$lib/profile.svelte";
   import { derivePromise, navigate, resolveLeafId } from "$lib/utils.svelte";
-  import { Category, Channel, Message } from "@roomy-chat/sdk";
+  import { Category, Channel } from "@roomy-chat/sdk";
   import toast from "svelte-french-toast";
   import { user } from "$lib/user.svelte";
   import { outerWidth } from "svelte/reactivity/window";
@@ -47,33 +45,6 @@
     }
   });
 
-  // TODO: track users via the space data
-  let users = derivePromise([], async () => {
-    if (!g.space) {
-      return [];
-    }
-
-    const result = new Set();
-    for (const channel of await g.space.channels.items()) {
-      for (const timelineItem of await channel.timeline.items()) {
-        const message = timelineItem.tryCast(Message);
-        if (message && message.authors.length > 0) {
-          for (const author of message.authors.toArray()) {
-            result.add(author);
-          }
-        }
-      }
-    }
-    const items = (await Promise.all(
-      [...result.values()].map(async (author) => {
-        const profile = await getProfile(author as string);
-        return { value: author, label: profile?.handle, category: "user" };
-      }),
-    )) as Item[];
-
-    return items;
-  });
-
   let availableThreads = derivePromise([], async () =>
     ((await g.space?.threads.items()) || []).filter((x) => !x.softDeleted),
   );
@@ -89,49 +60,6 @@
     if (!g.space) return [];
     return await g.space.sidebarItems.items();
   });
-
-  let contextItems: { value: Item[] } = derivePromise([], async () => {
-    if (!g.space) {
-      return [];
-    }
-    const items = [];
-
-    // add threads to list
-    for (const thread of await g.space.threads.items()) {
-      if (!thread.softDeleted) {
-        items.push({
-          value: JSON.stringify({
-            id: thread.id,
-            space: g.space.id,
-            type: "thread",
-          }),
-          label: thread.name,
-          category: "thread",
-        });
-      }
-    }
-
-    // add channels to list
-    items.push(
-      ...(await g.space.channels.items()).map((channel) => {
-        return {
-          value: JSON.stringify({
-            id: channel.id,
-            // TODO: I don't know that the space is necessary here or not.
-            space: g.space!.id,
-            type: "channel",
-          }),
-          label: channel.name,
-          category: "channel",
-        };
-      }),
-    );
-
-    return items;
-  });
-
-  setContext("users", users);
-  setContext("contextItems", contextItems);
 
   let showNewCategoryDialog = $state(false);
   let newCategoryName = $state("");
