+++ src/routes/(app)/[[spaceIndicator=dash]]/[space=space]/+layout.svelte
@@ -1,0 +1,81 @@
+<script lang="ts">
+  import { g } from '$lib/global.svelte';
+  import type { Item } from '$lib/tiptap/editor';
+  import { derivePromise } from '$lib/utils.svelte';
+  import { getProfile } from "$lib/profile.svelte";
+  import { setContext } from 'svelte';
+  import { Message } from '@roomy-chat/sdk';
+
+
+  // TODO: track users via the space data
+  let users = derivePromise([], async () => {
+    if (!g.space) {
+      return [];
+    }
+
+    const result = new Set();
+    for (const channel of await g.space.channels.items()) {
+      for (const timelineItem of await channel.timeline.items()) {
+        const message = timelineItem.tryCast(Message);
+        if (message && message.authors.length > 0) {
+          for (const author of message.authors.toArray()) {
+            result.add(author);
+          }
+        }
+      }
+    }
+    const items = (await Promise.all(
+      [...result.values()].map(async (author) => {
+        const profile = await getProfile(author as string);
+        return { value: author, label: profile?.handle, category: "user" };
+      }),
+    )) as Item[];
+
+    return items;
+  });
+  let contextItems: { value: Item[] } = derivePromise([], async () => {
+    if (!g.space) {
+      return [];
+    }
+    const items = [];
+
+    // add threads to list
+    for (const thread of await g.space.threads.items()) {
+      if (!thread.softDeleted) {
+        items.push({
+          value: JSON.stringify({
+            id: thread.id,
+            space: g.space.id,
+            type: "thread",
+          }),
+          label: thread.name,
+          category: "thread",
+        });
+      }
+    }
+
+    // add channels to list
+    items.push(
+      ...(await g.space.channels.items()).map((channel) => {
+        return {
+          value: JSON.stringify({
+            id: channel.id,
+            // TODO: I don't know that the space is necessary here or not.
+            space: g.space!.id,
+            type: "channel",
+          }),
+          label: channel.name,
+          category: "channel",
+        };
+      }),
+    );
+
+    return items;
+  });
+
+  setContext("users", users);
+  setContext("contextItems", contextItems);
+  let { children } = $props()
+</script>
+
+{@render children()}
